backend/requirements_hf_spaces.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.28.0
psycopg2-binary==2.9.7
pydantic==2.5.0
pydantic-settings==2.1.0
qdrant-client>=1.7.0
httpx==0.25.2
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
openai>=1.0.0
google-generativeai>=0.4.0
requests>=2.31.0
aiofiles>=23.0.0
PyJWT>=2.8.0
aiosqlite>=0.19.0
markdown>=3.4.0
beautifulsoup4>=4.12.0
pypdf>=3.0.0
tokenizers>=0.15.0
# Note: sentence-transformers excluded for Hugging Face Spaces
# Note: cohere excluded as not currently used